remove leftover websocket code
