--- components/common/BoardEditor/focalboard/src/components/newVersionBanner.scss
@@ -1,5 +1,0 @@
-.NewVersionBanner {
-    background-color: var(--prop-blue);
-    text-align: center;
-    padding: 10px;
-}
