+++ components/common/BoardEditor/focalboard/src/store/initialLoad.ts
@@ -1,11 +1,7 @@
 
-import { createAsyncThunk, createSelector } from '@reduxjs/toolkit';
+import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import charmClient from 'charmClient';
-
-import type { Subscription } from '../wsclient';
-
-import type { RootState } from './index';
 
 export const initialLoad = createAsyncThunk(
   'initialLoad',
@@ -29,11 +25,4 @@ export const initialReadOnlyLoad = createAsyncThunk(
     const blocks = charmClient.getSubtree(boardId, 3);
     return blocks;
   }
-);
-
-export const getUserBlockSubscriptions = (state: RootState): Subscription[] => state.users.blockSubscriptions;
-
-export const getUserBlockSubscriptionList = createSelector(
-  getUserBlockSubscriptions,
-  (subscriptions) => subscriptions
 );
