+++ package.json
@@ -21,10 +21,6 @@
     "server:prod": "node --require dd-trace/init node_modules/.bin/next start",
     "server:prod:win": "node --require dd-trace/init node_modules/next/dist/bin/next start",
     "server:staging": "npx react-env --path .env.staging -- node --require dd-trace/init node_modules/.bin/next start",
-    "sockets": "npm run sockets:dev",
-    "sockets:dev": "dotenv -e .env -e .env.local -- ts-node ./background/sockets.ts",
-    "sockets:prod": "node ./.next/server/sockets.js",
-    "sockets:staging": "npx react-env --path .env.staging -- node ./.next/server/sockets.js",
     "start": "npm run server:dev",
     "start:debug": "NODE_OPTIONS='--inspect' next",
     "start:dev:cron": "concurrently --names \"cron,next\" -c \"yellow,green\" \"npm run cron:dev\" \"npm run server:dev\"",
